requirements.txt
Flask==2.0.1
fastapi==0.68.0
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
tensorflow==2.6.0
torch==1.9.0
transformers==4.11.3
sqlalchemy==1.4.22
pydantic==1.8.2
requests==2.26.0
beautifulsoup4==4.10.0
pytest==6.2.4
notebook==6.4.3
docker==5.0.3
kubernetes==18.20.0
matplotlib==3.4.3
seaborn==0.11.2
plotly==5.1.0
streamlit==0.88.0
openpyxl==3.0.7
PyPDF2==1.26.0
python-docx==0.8.11
pytesseract==0.3.8
SpeechRecognition==3.8.1
faiss-cpu==1.7.4
neo4j==5.10.0